+++ sql/core/src/test/scala/org/apache/spark/sql/FileBasedDataSourceSuite.scala
@@ -33,7 +33,7 @@ import org.apache.spark.sql.catalyst.expressions.{AttributeReference, GreaterTha
 import org.apache.spark.sql.catalyst.expressions.IntegralLiteralTestUtils.{negativeInt, positiveInt}
 import org.apache.spark.sql.catalyst.plans.logical.Filter
 import org.apache.spark.sql.catalyst.types.DataTypeUtils
-import org.apache.spark.sql.execution.{FileSourceScanLike, SimpleMode}
+import org.apache.spark.sql.execution.{ExplainMode, FileSourceScanLike, SimpleMode}
 import org.apache.spark.sql.execution.adaptive.AdaptiveSparkPlanHelper
 import org.apache.spark.sql.execution.datasources.FilePartition
 import org.apache.spark.sql.execution.datasources.v2.{BatchScanExec, FileScan}
@@ -1243,6 +1243,50 @@ class FileBasedDataSourceSuite extends QueryTest
           }.flatten
           assert(filters.contains(GreaterThan(scan.logicalPlan.output.head, Literal(5L))))
         }
+      }
+    }
+  }
+
+  test("disable filter pushdown for collated strings") {
+    withTempPath { path =>
+      val collation = "'UCS_BASIC_LCASE'"
+      val df = sql(
+        s"""SELECT
+           |  COLLATE(c, $collation) as c1,
+           |  struct(COLLATE(c, $collation)) as str,
+           |  named_struct('f1', named_struct('f2', COLLATE(c, $collation), 'f3', 1)) as namedstr,
+           |  array(COLLATE(c, $collation)) as arr,
+           |  map(COLLATE(c, $collation), 1) as map1,
+           |  map(1, COLLATE(c, $collation)) as map2
+           |FROM VALUES ('aaa'), ('AAA'), ('bbb')
+           |as data(c)
+           |""".stripMargin)
+
+      df.write.parquet(path.getAbsolutePath)
+
+      // filter and expected result
+      val filters = Seq(
+        ("==", Seq(Row("aaa"), Row("AAA"))),
+        ("!=", Seq(Row("bbb"))),
+        ("<", Seq()),
+        ("<=", Seq(Row("aaa"), Row("AAA"))),
+        (">", Seq(Row("bbb"))),
+        (">=", Seq(Row("aaa"), Row("AAA"), Row("bbb"))))
+
+      filters.foreach { filter =>
+        val readback = spark.read
+          .parquet(path.getAbsolutePath)
+          .where(s"c1 ${filter._1} collate('aaa', $collation)")
+          .where(s"str ${filter._1} struct(collate('aaa', $collation))")
+          .where(s"namedstr.f1.f2 ${filter._1} collate('aaa', $collation)")
+          .where(s"arr ${filter._1} array(collate('aaa', $collation))")
+          .where(s"map_keys(map1) ${filter._1} array(collate('aaa', $collation))")
+          .where(s"map_values(map2) ${filter._1} array(collate('aaa', $collation))")
+          .select("c1")
+
+        val explain = readback.queryExecution.explainString(ExplainMode.fromString("extended"))
+        assert(explain.contains("PushedFilters: []"))
+        checkAnswer(readback, filter._2)
       }
     }
   }
